add : Mastery util함수 구현
